Crate rust_apl

This crate provides an interpreter for the language APL

This crate constructs an interpreter for the APL programming language, written in Rust. The primary inspiration is from Bob Nystrom’s excellent Crafting Interpreters. I also have used a significant amount of code from Emily Signet’s implementation of Lox.

The interpreter is relatively bare-bones. It supports a wide range of primitive glyphs, variables, and direct functions, but was created mostly as an instructive exercise for myself and is thus missing many features a full implementation would have. In this documentation I have tried to provide sufficient explanations for anyone who might want to try something similar.

Executable

See run.rs for the execution order and note how it is used for the interpreter executable. The below sections attempt to give a guide for each piece.

Scanning

First, we enumerate the different types of tokens with token::TokenType, a representation of the groups of characters that have a syntactic meaning in our language. These can be a single character, like ‘+’, or a group of characters, like ‘1.25’ or ‘hello’. This process is called scanning, performed in this crate by scanner::Scanner to produce a vector of token::Token.

TODO: put an example here

Parsing

Next, now that we have a series of token, we want to transform this into an expression. The key here is that any expression can be represented by a single tree like structure, called an abstact syntax tree. This process is called parsing, performed in this crate by parser::Parser to produce a vector of expr::Expr. Continuing with the previous example:

TODO: put an example

Evaluation

TODO

Variables

TODO

Functions

TODO

Modules

apl_type 🔒

environment 🔒

err

expr 🔒

interpreter

parser

primitives 🔒

run

scanner

token 🔒